requirements.txt
Django==1.6.5
PyYAML==3.11
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-toolbelt==0.0.1
gunicorn==19.0.0
psycopg2==2.5.3
pystache==0.5.4
static==1.0.2
wsgiref==0.1.2
